+++ test/btcrelay_fee.se
@@ -1,0 +1,4 @@
+# macro OFFSET_ABI: 68 + 32  # extra 32 bytes since storeBlockWithFee has extra param
+
+# testing overrides should be before the contract being tested
+inset('../btcrelay.se')
